requirements.txt
   python-telegram-bot==20.0
   selenium==4.0.0
   webdriver-manager==3.5.0